remove unneccessary code by using flask_login current_user and blueprint classes
